requirements.txt
streamlit==1.32.0
openai==1.12.0
numpy==1.26.0
python-dotenv==1.0.0
boto3==1.36.17
request
# SpeechRecognition
# playsound==1.2.2
# pipwin
# pyaudio